refactor(model): 정확한 수익률 계산을 위해 float -> BigDecimal 로 변경

+++ src/main/java/lotto/dto/LottoTotalReturnDTO.java
@@ -1,12 +1,13 @@
 package lotto.dto;
 
+import java.math.BigDecimal;
 import java.text.DecimalFormat;
 import lotto.model.LottoRankResult;
 
 public record LottoTotalReturnDTO(LottoRankResult result, int countOfLotto) {
     public String getTotalReturn() {
         DecimalFormat decimalFormat = new DecimalFormat("###,###.#");
-        float totalReturn = result.getTotalReturn(countOfLotto);
+        BigDecimal totalReturn = result.getTotalReturn(countOfLotto);
         return decimalFormat.format(totalReturn);
     }
 }
